chore(homepage): move lastDeployedResources to their own folder (#2860)
