+++ apps/frontend/src/components/homepage/PrivateHomepage.tsx
@@ -1,6 +1,6 @@
 import BlueBlurDecoration from '@/components/homepage/BlueBlurDecoration';
 import { buildDistinctPlatformIdentifiersFromServiceDefinition } from '@/components/homepage/Homepage.utils';
-import LastDeployedResourcesSection from '@/components/homepage/LastDeployedResourcesSection';
+import LastDeployedResourcesSection from '@/components/homepage/last-deployed-resources/LastDeployedResourcesSection';
 import { RegisteredPlatformsSection } from '@/components/homepage/registered-platforms/RegisteredPlatformsSection';
 import MostDeployedResources from '@/components/homepage/resources/MostDeployedResources';
 import NewestResources from '@/components/homepage/resources/NewestResources';
